+++ tests/Frontend/components/Room/RoomFile.spec.js
@@ -1,7 +1,7 @@
 import { mount } from '@vue/test-utils';
 import BootstrapVue, { BButton, BFormFile, BFormInvalidFeedback, BModal, BTbody } from 'bootstrap-vue';
 import FileComponent from '../../../../resources/js/components/Room/FileComponent.vue';
-import VueClipboard from 'vue-clipboard2';
+
 import Base from '../../../../resources/js/api/base';
 import PermissionService from '../../../../resources/js/services/PermissionService';
 import _ from 'lodash';
@@ -11,8 +11,7 @@ import { createTestingPinia } from '@pinia/testing';
 
 const localVue = createLocalVue();
 
-localVue.use(BootstrapVue);
-localVue.use(VueClipboard);
+localVue.use(BootstrapVue)
 localVue.use(PiniaVuePlugin);
 
 const exampleUser = { id: 1, firstname: 'John', lastname: 'Doe', locale: 'de', permissions: ['rooms.create'], model_name: 'User', room_limit: -1 };