+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Layout issue in access for participants ([#653], [#654])
 
 ## [v2.3.1] - 2023-10-16
 ### Changed
@@ -669,6 +671,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#586]: https://github.com/THM-Health/PILOS/pull/586
 [#620]: https://github.com/THM-Health/PILOS/issues/620
 [#621]: https://github.com/THM-Health/PILOS/pull/621
+[#653]: https://github.com/THM-Health/PILOS/issues/653
+[#654]: https://github.com/THM-Health/PILOS/pull/654
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v2.3.1...develop
 [v1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0