+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Custom create parameters in room type settings ([#574])
+
+### Changed
+- The learning dashboard is no longer always disabled and the meeting layout is no longer always "custom". These settings can be overridden with the room type settings. If the two settings are not set in the room type settings, the previous values apply until the next release; afterwards, these settings can only be set via the room type settings. ([#574])
 
 ## [v2.3.1] - 2023-10-16
 ### Changed