+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v2.3.2] - 2023-11-06
 ### Fixed
 - Layout issue in access for participants ([#653], [#654])
 
@@ -677,7 +679,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#653]: https://github.com/THM-Health/PILOS/issues/653
 [#654]: https://github.com/THM-Health/PILOS/pull/654
 
-[unreleased]: https://github.com/THM-Health/PILOS/compare/v2.3.1...develop
+[unreleased]: https://github.com/THM-Health/PILOS/compare/v2.3.2...develop
 [v1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0
 [v1.1.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.1.0
 [v1.2.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.2.0
@@ -710,3 +712,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v2.2.0]: https://github.com/THM-Health/PILOS/releases/tag/v2.2.0
 [v2.3.0]: https://github.com/THM-Health/PILOS/releases/tag/v2.3.0
 [v2.3.1]: https://github.com/THM-Health/PILOS/releases/tag/v2.3.1
+[v2.3.2]: https://github.com/THM-Health/PILOS/releases/tag/v2.3.2